pyproject.toml
[project]
name = "count-tokens"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "pathspec>=1.0.4",
    "python-dotenv>=1.2.2",
    "questionary>=2.1.1",
    "rich>=14.3.3",
    "tomli-w>=1.2.0",
    "typer>=0.24.1",
]

[project.scripts]
count-tokens = "count_tokens.cli:app"
ct = "count_tokens.cli:app"

[tool.pytest.ini_options]
markers = [
    "claude: tests requiring Anthropic API key",
    "openai: tests requiring OpenAI API key",
    "gemini: tests requiring Gemini API key",
    "grok: tests requiring xAI API key",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
